fix alloc_aligned

+++ case_studies/pic/include/optitrust.h
@@ -39,13 +39,13 @@ void* MMALLOC3(int N1, int N2, int N3, size_t bytes_per_item);
 void* MMALLOC4(int N1, int N2, int N3, int N4, size_t bytes_per_item);
 
 
-void* MMALLOC_ALIGNED1(int N1, size_t bytes_per_item, size_t alignment);
+void* MMALLOC_ALIGNED1(size_t N1, size_t bytes_per_item, size_t alignment);
 
-void* MMALLOC_ALIGNED2(int N1, int N2, size_t bytes_per_item, size_t alignment);
+void* MMALLOC_ALIGNED2(size_t N1, size_t N2, size_t bytes_per_item, size_t alignment);
 
-void* MMALLOC_ALIGNED3(int N1, int N2, int N3, size_t bytes_per_item, size_t alignment);
+void* MMALLOC_ALIGNED3(size_t N1, size_t N2, size_t N3, size_t bytes_per_item, size_t alignment);
 
-void* MMALLOC_ALIGNED4(int N1, int N2, int N3, int N4, size_t bytes_per_item, size_t alignment);
+void* MMALLOC_ALIGNED4(size_t N1, size_t N2, size_t N3, size_t N4, size_t bytes_per_item, size_t alignment);
 
 void MFREE(void* p);
 void MFREE1(int N1, void* p);
